requirements.txt
deta==1.2.0
matplotlib==3.9.0
python-dotenv==1.0.1
requests==2.32.2
schedule==1.2.1
seaborn==0.13.2
streamlit==1.35.0
pandas==2.2.2
calmap==0.0.11
plotly==5.22.0